refactor: split LocationHomepagePage into hook and block renderer
678 lines held the homepage query, six mutations with optimistic cache
updates, slot-key bookkeeping, a ~250-line per-block editor dispatch, and
the page chrome.

Split into:
- useLocationHomepageEditor.ts        query + mutations + state
- LocationHomepageBlockRenderer.tsx   per-block editor dispatch + wiring
- locationHomepagePage.utils.ts       getLocationLabel
- LocationHomepagePage.tsx            page chrome (678 -> 237 lines)

One subtlety this refactor had to get right: each editor carried
key={homepageBlockEditorIdentity(block).join(':')}, and that key is load-
bearing. HomepageBlocksSortableList keys its wrappers by block.id alone, so
after a convert (same id, new blockType/slotCount) the mounted editor and
its TanStack cache entry would be reused. Returning the editor from a child
component would have moved that key to the component's root element, where
it no longer forces a remount.

So the key now sits on <LocationHomepageBlockRenderer> itself, via an
exported locationHomepageBlockKey() that mirrors the original branch order
exactly — including that unknown blocks, and location-grid blocks with no
child level, must key on block.id, since homepageBlockEditorIdentity reads
selection.totalSlots and UnknownBlockResponse has no selection.

Verified: tsc error signature identical to baseline; check-boundaries output
identical to main; vitest 333 pass / 3 pre-existing failures, with all 20
homepageFeaturedContent tests passing.
